Move install inside commands.
