+++ hatch/commands/__init__.py
@@ -3,3 +3,4 @@ from hatch.commands.python import python
 from hatch.commands.config import config
 from hatch.commands.new import new
 from hatch.commands.init import init
+from hatch.commands.install import install
